requirements.txt
google-adk
google-genai
vertexai
streamlit
uvicorn
fastapi
google-cloud-storage
python-dotenv
gunicorn
google-api-python-client
google-auth
requests